pyproject.toml
[project]
name = "fastapi-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "bcrypt>=4.3.0",
    "cryptography>=45.0.5",
    "email-validator>=2.2.0",
    "fastapi>=0.115.14",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
